move config and remove no op (we always need the chain id now)
