+++ evmd/config_testing.go
@@ -5,6 +5,7 @@ package evmd
 
 import (
 	"fmt"
+	"github.com/cosmos/evm/cmd/evmd/config"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
 
 	"cosmossdk.io/math"
@@ -16,40 +17,40 @@ import (
 // that allows initializing the app with different coin info based on the
 // chain id
 var ChainsCoinInfo = map[uint64]evmtypes.EvmCoinInfo{
-	EighteenDecimalsChainID: {
-		Denom:         ExampleChainDenom,
-		ExtendedDenom: ExampleChainDenom,
-		DisplayDenom:  ExampleDisplayDenom,
+	config.EighteenDecimalsChainID: {
+		Denom:         config.ExampleChainDenom,
+		ExtendedDenom: config.ExampleChainDenom,
+		DisplayDenom:  config.ExampleDisplayDenom,
 		Decimals:      evmtypes.EighteenDecimals,
 	},
-	SixDecimalsChainID: {
+	config.SixDecimalsChainID: {
 		Denom:         "utest",
 		ExtendedDenom: "atest",
 		DisplayDenom:  "test",
 		Decimals:      evmtypes.SixDecimals,
 	},
-	TwelveDecimalsChainID: {
+	config.TwelveDecimalsChainID: {
 		Denom:         "ptest2",
 		ExtendedDenom: "atest2",
 		DisplayDenom:  "test2",
 		Decimals:      evmtypes.TwelveDecimals,
 	},
-	TwoDecimalsChainID: {
+	config.TwoDecimalsChainID: {
 		Denom:         "ctest3",
 		ExtendedDenom: "atest3",
 		DisplayDenom:  "test3",
 		Decimals:      evmtypes.TwoDecimals,
 	},
-	TestChainID1: {
-		Denom:         ExampleChainDenom,
-		ExtendedDenom: ExampleChainDenom,
-		DisplayDenom:  ExampleChainDenom,
+	config.TestChainID1: {
+		Denom:         config.ExampleChainDenom,
+		ExtendedDenom: config.ExampleChainDenom,
+		DisplayDenom:  config.ExampleChainDenom,
 		Decimals:      evmtypes.EighteenDecimals,
 	},
-	TestChainID2: {
-		Denom:         ExampleChainDenom,
-		ExtendedDenom: ExampleChainDenom,
-		DisplayDenom:  ExampleChainDenom,
+	config.TestChainID2: {
+		Denom:         config.ExampleChainDenom,
+		ExtendedDenom: config.ExampleChainDenom,
+		DisplayDenom:  config.ExampleChainDenom,
 		Decimals:      evmtypes.EighteenDecimals,
 	},
 }
