+++ testutil/constants/constants_test.go
@@ -6,14 +6,13 @@ import (
 	"github.com/stretchr/testify/require"
 
 	chainconfig "github.com/cosmos/evm/cmd/evmd/config"
-	"github.com/cosmos/evm/evmd"
 	"github.com/cosmos/evm/testutil/constants"
 )
 
 func TestRequireSameTestDenom(t *testing.T) {
 	require.Equal(t,
 		constants.ExampleAttoDenom,
-		evmd.ExampleChainDenom,
+		chainconfig.ExampleChainDenom,
 		"test denoms should be the same across the repo",
 	)
 }
@@ -29,7 +28,7 @@ func TestRequireSameTestBech32Prefix(t *testing.T) {
 func TestRequireSameWEVMOSMainnet(t *testing.T) {
 	require.Equal(t,
 		constants.WEVMOSContractMainnet,
-		evmd.WEVMOSContractMainnet,
+		chainconfig.WEVMOSContractMainnet,
 		"wevmos contract addresses should be the same across the repo",
 	)
 }
